Move Emitter- and TrackerDefaults to core
